requirements.txt
django
openai==1.35.9
pymongo==4.8.0
google-generativeai==0.7.1